+++ __tests__/getTestRule.test.js
@@ -8,11 +8,14 @@ testRule({
 	plugins: [require.resolve('./fixtures/plugin-foo.js')],
 	ruleName: 'plugin/foo',
 	config: ['.a'],
-	invalidConfig: [123],
 
 	accept: [
 		{
 			code: '.a {}',
+		},
+		{
+			code: '.a {}',
+			description: 'with description',
 		},
 	],
 
@@ -20,6 +23,11 @@ testRule({
 		{
 			code: '#a {}',
 			message: 'No "#a" selector (plugin/foo)',
+		},
+		{
+			code: '#a {}',
+			message: 'No "#a" selector (plugin/foo)',
+			description: 'with description',
 		},
 	],
 });