+++ jest-setup.js
@@ -1,5 +1,7 @@
 'use strict';
 
-const getTestRule = require('./getTestRule');
+const getTestRule = require('./getTestRule.js');
+const getTestInvalidRuleConfigs = require('./getTestInvalidRuleConfigs.js');
 
 global.testRule = getTestRule();
+global.testInvalidRuleConfigs = getTestInvalidRuleConfigs();